refactor CLIHandler for password handling, updated project structure
