+++ .github/workflows/maven.yml
@@ -10,14 +10,14 @@ jobs:
   build:
     runs-on: ubuntu-latest
     permissions:
-      contents: write    # Allows the action to write to the dependency graph
-      security-events: write  # Allows the action to record security-related events
+      contents: write  
+      security-events: write  
 
     steps:
     - name: Checkout Repository
       uses: actions/checkout@v4
       with:
-        token: ${{ secrets.PERSONAL_ACCESS_TOKEN }}  # Using PAT for authentication if necessary
+        token: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
 
     - name: Set up JDK 21
       uses: actions/setup-java@v4
@@ -26,17 +26,10 @@ jobs:
         distribution: 'temurin'
         cache: maven
 
-    - name: Build and Test with Maven
-      run: mvn clean verify
-
-    - name: Upload JAR Artifact
-      uses: actions/upload-artifact@v4
-      with:
-        name: Alohomora-JAR
-        path: target/*.jar
-        retention-days: 7
+    - name: Test with Maven
+      run: mvn test
 
     - name: Update Dependency Graph
       uses: advanced-security/maven-dependency-submission-action@v4
       with:
-        token: ${{ secrets.PERSONAL_ACCESS_TOKEN }}  # Ensure the PAT is used for submitting the snapshot
+        token: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
