+++ .github/workflows/release.yml
@@ -24,7 +24,7 @@ jobs:
 
       - name: Rename JAR for Release
         run: |
-          mv target/*-app.jar target/Alohomora-${{ github.ref_name }}-app.jar
+          mv target/*-client.jar target/Alohomora-${{ github.ref_name }}-client.jar
           mv target/*-background.jar target/Alohomora-${{ github.ref_name }}-background.jar
 
       - name: Create GitHub Release and Upload JAR
@@ -32,9 +32,11 @@ jobs:
         with:
           tag_name: ${{ github.ref_name }}
           name: Release ${{ github.ref_name }}
-          body:
-                This is the release of Alohomora version ${{ github.ref_name }}.
-                The JAR file is attached below.
-          files: target/Alohomora-${{ github.ref_name }}.jar
+          body: |
+            This is the release of Alohomora version ${{ github.ref_name }}.
+            The JAR files are attached below. (client and background)
+          files: |
+            target/Alohomora-${{ github.ref_name }}-client.jar
+            target/Alohomora-${{ github.ref_name }}-background.jar
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
