+++ src/main/java/io/dayfit/github/shared/JSON.java
@@ -1,4 +1,4 @@
-package io.dayfit.github.backgroundServices.utils;
+package io.dayfit.github.shared;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
