+++ src/main/resources/application.properties
@@ -1,4 +1,11 @@
 application.version=@application.version@
 max.wait.time.backgroundServices=@max.wait.time.backgroundServices@
+
+//Spring server related properties
+
+logging.level.root=WARN
+logging.level.org.springframework=WARN
+
 server.ping.response=@server.ping.response@
+spring.main.banner-mode=off
 server.port=8080
