+++ src/test/java/io/dayfit/github/shared/JSONTest.java
@@ -1,5 +1,6 @@
-package io.dayfit.github.backgroundServices.utils;
+package io.dayfit.github.shared;
 import com.fasterxml.jackson.core.JsonProcessingException;
+
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
